+++ tests/test_common/test_storage/test_lmdb.py
@@ -9,6 +9,7 @@ Minos framework can not be copied and/or distributed without the express permiss
 import shutil
 
 import pytest
+
 from minos.common import (
     MinosStorageLmdb,
 )